20220625 장애물 회피를 위한 기본함수 추가, 기존 알고리즘 코드 제거

+++ src/lib/cog_predict_collision.h
@@ -4,5 +4,6 @@
 
 #include "PosInfo.h"
 #include "LineEq.h"
+#include "CircEq.h"
 
-float cog_predict_collision(std::vector<PosInfo> obstacles, LineEq line);
+int cog_predict_collision(std::vector<PosInfo> obstacles, LineEq line);
